fix(link): Override default behavior for opening a link with ctrl (meta) click or alt-enter.
While the default behavior of CKEditor (Open in a new browser tab) is exactly what we need for external links it is not what we want for content links. Content Links should open in the same way it is opened from the link balloon (new work area tab).
