+++ packages/ckeditor5-coremedia-link/src/contentlink/ContentLinks.ts
@@ -13,7 +13,7 @@ import LinkFormView from "@ckeditor/ckeditor5-link/src/ui/linkformview";
 import "../lang/contentlink";
 import ContentLinkClipboardPlugin from "./ContentLinkClipboardPlugin";
 import { OpenInTabCommand } from "@coremedia/ckeditor5-coremedia-content/commands/OpenInTabCommand";
-import LinkReadOnlyPlugin from "./LinkReadOnlyPlugin";
+import LinkUserActionsPlugin from "./LinkUserActionsPlugin";
 
 /**
  * This plugin allows content objects to be dropped into the link dialog.
@@ -28,7 +28,7 @@ export default class ContentLinks extends Plugin {
     ContentLinkFormViewExtension,
     ContentLinkCommandHook,
     ContentLinkClipboardPlugin,
-    LinkReadOnlyPlugin,
+    LinkUserActionsPlugin,
   ];
 
   init(): void {
