+++ packages/ckeditor5-coremedia-link/src/contentlink/LinkUserActionsPlugin.ts
@@ -1,0 +1,179 @@
+import Plugin from "@ckeditor/ckeditor5-core/src/plugin";
+import Editor from "@ckeditor/ckeditor5-core/src/editor/editor";
+import View from "@ckeditor/ckeditor5-engine/src/view/view";
+import ViewDocumentFragment from "@ckeditor/ckeditor5-engine/src/view/documentfragment";
+import ViewElement from "@ckeditor/ckeditor5-engine/src/view/element";
+import { serviceAgent } from "@coremedia/service-agent";
+import { createWorkAreaServiceDescriptor } from "@coremedia/ckeditor5-coremedia-studio-integration/content/WorkAreaServiceDescriptor";
+import { openLink } from "@ckeditor/ckeditor5-link/src/utils";
+import LoggerProvider from "@coremedia/ckeditor5-logging/logging/LoggerProvider";
+import Logger from "@coremedia/ckeditor5-logging/logging/Logger";
+import TextProxy from "@ckeditor/ckeditor5-engine/src/model/textproxy";
+import ModelElement from "@ckeditor/ckeditor5-engine/src/model/element";
+import ModelNode from "@ckeditor/ckeditor5-engine/src/model/node";
+import {
+  isModelUriPath,
+  requireContentUriPath,
+} from "@coremedia/ckeditor5-coremedia-studio-integration/content/UriPath";
+import { env, keyCodes } from "@ckeditor/ckeditor5-utils";
+
+export default class LinkUserActionsPlugin extends Plugin {
+  static readonly pluginName: string = "LinkUserActionsPlugin";
+  static readonly LOG: Logger = LoggerProvider.getLogger(LinkUserActionsPlugin.pluginName);
+  static readonly requires = [];
+
+  init(): void {
+    this.#handleLinkClicksInReadOnly();
+    this.#preventOpenInNewTabForContentLinks();
+  }
+
+  #handleLinkClicksInReadOnly() {
+    const editor = this.editor;
+    const view = editor.editing.view;
+    const viewDocument = view.document;
+
+    this.listenTo(
+      viewDocument,
+      "click",
+      (evt, data: { domTarget: Element; preventDefault: () => void; view: View }) => {
+        if (!editor.isReadOnly) {
+          return;
+        }
+        let clickedElement: Element | null = (data as { domTarget: Element }).domTarget;
+        if (clickedElement.tagName.toLowerCase() !== "a") {
+          clickedElement = clickedElement.closest("a");
+        }
+        if (!clickedElement) {
+          return;
+        }
+
+        evt.stop();
+        data.preventDefault();
+        this.#onReadOnlyLinkActivated(editor, data.view, clickedElement);
+      },
+      //@ts-expect-error context is not part of the types but in ckeditor5-link/src/linkediting the event is caught in capture phase
+      { priority: "high", context: "$capture" }
+    );
+  }
+
+  #onReadOnlyLinkActivated(editor: Editor, view: View, domElement: Element): void {
+    const modelElement: TextProxy | undefined = this.#resolveAnchorModelElement(editor, view, domElement);
+    if (!modelElement) {
+      return;
+    }
+    const linkHref: string = modelElement.getAttribute("linkHref") as string;
+    if (isModelUriPath(linkHref)) {
+      this.#openInWorkAreaTab(linkHref);
+    } else {
+      openLink(linkHref);
+    }
+  }
+
+  #preventOpenInNewTabForContentLinks() {
+    const editor = this.editor;
+    const view = editor.editing.view;
+    const viewDocument = view.document;
+
+    this.listenTo(
+      viewDocument,
+      "click",
+      (
+        evt,
+        data: {
+          domEvent: { metaKey: boolean; ctrlKey: boolean };
+          domTarget: Element;
+          preventDefault: () => void;
+          view: View;
+        }
+      ) => {
+        if (editor.isReadOnly) {
+          return;
+        }
+        const shouldOpen: boolean = env.isMac ? data.domEvent.metaKey : data.domEvent.ctrlKey;
+        if (!shouldOpen) {
+          return;
+        }
+
+        let clickedElement: Element | null = (data as { domTarget: Element }).domTarget;
+        if (clickedElement.tagName.toLowerCase() !== "a") {
+          clickedElement = clickedElement.closest("a");
+        }
+        if (!clickedElement) {
+          return;
+        }
+
+        const anchorModelElement = this.#resolveAnchorModelElement(editor, view, clickedElement);
+        if (!anchorModelElement) {
+          return;
+        }
+
+        if (anchorModelElement.hasAttribute("linkHref")) {
+          const linkHref = anchorModelElement.getAttribute("linkHref");
+          if (isModelUriPath(linkHref)) {
+            evt.stop();
+            data.preventDefault();
+            this.#openInWorkAreaTab(requireContentUriPath(linkHref));
+          }
+        }
+      },
+      //@ts-expect-error context is not part of the types but in ckeditor5-link/src/linkediting the event is caught in capture phase
+      { priority: "high", context: "$capture" }
+    );
+
+    this.listenTo(
+      viewDocument,
+      "keydown",
+      (evt, data: { keyCode: number; altKey: boolean }) => {
+        if (editor.isReadOnly) {
+          return;
+        }
+        const url = editor.commands?.get("link")?.value;
+        if (!url) {
+          return;
+        }
+        const shouldOpen: boolean = data.keyCode === keyCodes.enter && data.altKey;
+        if (!shouldOpen) {
+          return;
+        }
+        if (isModelUriPath(url)) {
+          evt.stop();
+          this.#openInWorkAreaTab(requireContentUriPath(url));
+        }
+      },
+      { priority: "high" }
+    );
+  }
+
+  #resolveAnchorModelElement(editor: Editor, view: View, domElement: Element): TextProxy | undefined {
+    //@ts-expect-error bad typings, mapDomToView parameter is typed as model.element, but it should be the typescript element.
+    const viewElement: ViewElement | ViewDocumentFragment | undefined = view.domConverter.mapDomToView(domElement);
+    if (!viewElement || viewElement instanceof ViewDocumentFragment) {
+      return;
+    }
+    const viewRange = view.createRangeIn(viewElement);
+    const modelRange = editor.editing.mapper.toModelRange(viewRange);
+    const modelItemsInRange = Array.from(modelRange.getItems());
+    const textProxies: TextProxy[] = modelItemsInRange
+      .filter((item: ModelNode | ModelElement | Text | TextProxy) => item instanceof TextProxy)
+      .map((textProxy) => textProxy as TextProxy)
+      .filter((textProxy) => textProxy.hasAttribute("linkHref"));
+    if (textProxies.length < 1) {
+      LinkUserActionsPlugin.LOG.debug("No links found after click");
+      return undefined;
+    }
+    if (textProxies.length > 1) {
+      LinkUserActionsPlugin.LOG.debug("Found multiple clicked links, taking the first to open.");
+    }
+
+    return textProxies[0];
+  }
+
+  #openInWorkAreaTab(uri: string): void {
+    serviceAgent
+      .fetchService(createWorkAreaServiceDescriptor())
+      .then((workAreaService) => workAreaService.openEntitiesInTabs([requireContentUriPath(uri)]))
+      .catch((reason) => {
+        LinkUserActionsPlugin.LOG.warn(reason);
+      });
+  }
+}
